Added Grunt build process for iOS.
Integrated configuration and a custom FeedHenry platform specific
Grunt task to build for the iOS PhoneGap project, allowing us to then
open the builds/ios folder in Xcode to run in the simulator or build
etc.

Added package containing the cordova js lib as well as an index.html
which includes reference to that file, which is included within an
iOS specific build.

Also renamed cordova_projects to builds for clarity.