Adjusted fhbuild and added thorough ios cleanup.
Fixed a few potential problems with the fhbuild script, by adding
randomised IDs to the targets we create, in order to prevent any
conflicts with pre-existing targets we want in the config.

Also moved require.js file to app dir in order to facilitate the
clean and adjusted refs in index.html files accordingly.