Replace Zepto with jQuery.
The decision to use Zepto for mobile and jQuery for desktop variants
of the app was premature; since doing more research, it appears that
Zepto, while smaller in size, actually performs vastly worse than
standard jQuery in most cases. File size isn't even a consideration
for the packaged app.

As well as this, the use of a jQuery calendar plugin which doesn't
work with Zepto is another factor. It's simply easier across the
board to use jQuery.