Cleanup & Cordova dependant pushState setting.
Support for the pushState API in Backbone was previously just
disabled, however we need to create a backend which is as platform
agnostic as possible, in order to save ourselves large amounts of
trouble in future. The app.js file should now be completely ready for
this, as no part of it is platform dependant.